+++ src/main/java/one/util/streamex/AbstractStreamEx.java
@@ -219,6 +219,25 @@ public abstract class AbstractStreamEx<T, S extends AbstractStreamEx<T, S>> exte
         return new DoubleStreamEx(stream().flatMapToDouble(mapper), context);
     }
 
+    /**
+     * Returns a new stream containing all the elements of the original stream interspersed with
+     * given delimiter.
+     * 
+     * <p>
+     * For example, {@code StreamEx.of("a", "b", "c").intersperse("x")} will yield a stream containing
+     * five elements: a, x, b, x, c.
+     * 
+     * <p>
+     * This is an <a href="package-summary.html#StreamOps">intermediate operation</a>.
+     * 
+     * @param delimiter a delimiter to be inserted between each pair of elements
+     * @return the new stream
+     * @since 0.6.6
+     */
+    public S intersperse(T delimiter) {
+        return supply(stream().flatMap(s -> StreamEx.of(delimiter, s)).skip(1));
+    }
+
     @Override
     public S distinct() {
         return supply(stream().distinct());