+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 Check also [MIGRATION.md](MIGRATION.md) for possible compatibility problems.
 
+### 0.6.6
+* [#145] Added: `intersperse` method for all stream types.
+
 ### 0.6.5
 * [#137] Added: `StreamEx.toNavigableMap()`, `EntryStream.toNavigableMap()`
 * [#141] Added: `StreamEx.flatArray()`, `EntryStream.flatArray()`